Sloppy add of Options, but entire flow is working. Possibly a cleaner way then what is here...

+++ src/main/scala/akka/stream/alpakka/sqs/newdsl/SqsAckFlow.scala
@@ -1,11 +1,12 @@
 package akka.stream.alpakka.sqs.newdsl
 
 import akka.NotUsed
-import akka.stream.scaladsl.{Flow}
+import akka.stream.scaladsl.Flow
 import software.amazon.awssdk.services.sqs.SqsAsyncClient
 import software.amazon.awssdk.services.sqs.model.DeleteMessageRequest
 
 import scala.compat.java8.FutureConverters._
+import scala.concurrent.{ExecutionContext, Future}
 
 /**
   * Author: Nicholas Connor
@@ -29,16 +30,20 @@ object SqsAckFlow {
       }
 
   def ignored(parallelism: Int)
-           (implicit sqsAsyncClient: SqsAsyncClient): Flow[MessageAction, ActionResult, NotUsed] =
-    Flow[MessageAction]
+           (implicit sqsAsyncClient: SqsAsyncClient, ec: ExecutionContext): Flow[Option[MessageAction], ActionResult, NotUsed] =
+    Flow[Option[MessageAction]]
       .mapAsyncUnordered(parallelism) {
-        case delete: MessageAction.Delete =>
-          sqsAsyncClient.deleteMessage(
-            DeleteMessageRequest.builder()
-              .queueUrl(delete.queueUrl)
-              .receiptHandle(delete.message.receiptHandle)
-              .build()
-          ).thenApply[ActionResult](_ => ActionResult.Ignored)
-           .toScala
-      }
+            case delete: Option[MessageAction.Delete] =>
+              if (delete.nonEmpty) {
+                sqsAsyncClient.deleteMessage(
+                  DeleteMessageRequest.builder()
+                    .queueUrl(delete.get.queueUrl)
+                    .receiptHandle(delete.get.message.receiptHandle)
+                    .build()
+                ).thenApply[ActionResult](_ => ActionResult.Ignored)
+                  .toScala
+              } else {
+                Future { ActionResult.Ignored }
+              }
+          }
 }
